Move TdsParserStateObject and multiplexer partial Move is made wholesale to preserve history
