+++ src/Microsoft.Data.SqlClient/tests/FunctionalTests/Microsoft.Data.SqlClient.FunctionalTests.csproj
@@ -21,7 +21,7 @@
     <!-- should ideally be UnitTests if they need to access internals. But detangling these    -->
     <!-- tests is far beyond the scope of the common project, so it is left as a TODO for now. -->
     <Compile Include="$(RepoRoot)src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/Packet.cs" />
-    <Compile Include="$(RepoRoot)src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/TdsParserStateObject.Multiplexer.cs" />
+    <Compile Include="$(RepoRoot)src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/Parser/TdsParserStateObject.Multiplexer.cs" />
   </ItemGroup>
 
   <!-- Embedded resources and content files  =========================== -->
